Most of FX rack param section
Worked out most of the parameters for fx. Types of fx need to be worked out; does changing it determine what DSP is used internally, since there are several types used for reverb, but most everything else (flanger/phaser/etc) just has one "type" used by all of them? There also seem to be some params that are not exposed in the UI (and are mostly consistent regardless of which fx is loaded). Probably the only way to answer some of these questions will be to create a custom library with fx settings and type combinations that don't ever occur in normal use (potential to be risky experiment).